Moved WooLogViewerScreen to it's own file
